# Sync with go1.23