Replay of https://github.com/kubernetes-sigs/json/pull/19 after resolving https://github.com/golang/go/issues/69828

First commit adjusts benchmark data to contain a variety of slice sizes to give a better representation of general performance

Second and third commits are picked from https://github.com/kubernetes-sigs/json/pull/19

Benchmark results (using the new benchmark data, before and after the code changes):
```
                                   │   old.txt   │              new.txt               │
                                   │   sec/op    │   sec/op     vs base               │
Unmarshal/typed_stdlib-10            37.10µ ± 4%   37.82µ ± 1%  +1.95% (p=0.023 n=10)
Unmarshal/typed_unmarshal-10         38.02µ ± 1%   37.54µ ± 6%       ~ (p=0.143 n=10)
Unmarshal/typed_strict-10            39.43µ ± 0%   38.66µ ± 1%  -1.95% (p=0.000 n=10)
Unmarshal/typed_strict-custom-10     39.74µ ± 1%   39.13µ ± 1%  -1.53% (p=0.004 n=10)
Unmarshal/untyped_stdlib-10          36.77µ ± 2%   37.97µ ± 2%  +3.27% (p=0.001 n=10)
Unmarshal/untyped_unmarshal-10       36.88µ ± 1%   36.47µ ± 1%  -1.09% (p=0.004 n=10)
Unmarshal/untyped_strict-10          41.17µ ± 1%   40.43µ ± 5%       ~ (p=0.089 n=10)
Unmarshal/untyped_strict-custom-10   41.48µ ± 2%   40.60µ ± 2%  -2.11% (p=0.023 n=10)
geomean                              38.78µ        38.56µ       -0.58%

                                   │   old.txt    │                new.txt                 │
                                   │     B/op     │     B/op      vs base                  │
Unmarshal/typed_stdlib-10            63.38Ki ± 0%   63.38Ki ± 0%        ~ (p=1.000 n=10) ¹
Unmarshal/typed_unmarshal-10         74.17Ki ± 0%   63.44Ki ± 0%  -14.47% (p=0.000 n=10)
Unmarshal/typed_strict-10            75.02Ki ± 0%   64.28Ki ± 0%  -14.31% (p=0.000 n=10)
Unmarshal/typed_strict-custom-10     75.05Ki ± 0%   64.31Ki ± 0%  -14.30% (p=0.000 n=10)
Unmarshal/untyped_stdlib-10          28.30Ki ± 0%   28.30Ki ± 0%        ~ (p=0.466 n=10)
Unmarshal/untyped_unmarshal-10       28.57Ki ± 0%   28.38Ki ± 0%   -0.65% (p=0.000 n=10)
Unmarshal/untyped_strict-10          29.13Ki ± 0%   28.94Ki ± 0%   -0.64% (p=0.000 n=10)
Unmarshal/untyped_strict-custom-10   29.16Ki ± 0%   28.97Ki ± 0%   -0.65% (p=0.000 n=10)
geomean                              45.44Ki        42.77Ki        -5.88%
¹ all samples are equal

                                   │  old.txt   │               new.txt               │
                                   │ allocs/op  │ allocs/op   vs base                 │
Unmarshal/typed_stdlib-10            166.0 ± 0%   166.0 ± 0%       ~ (p=1.000 n=10) ¹
Unmarshal/typed_unmarshal-10         172.0 ± 0%   166.0 ± 0%  -3.49% (p=0.000 n=10)
Unmarshal/typed_strict-10            182.0 ± 0%   176.0 ± 0%  -3.30% (p=0.000 n=10)
Unmarshal/typed_strict-custom-10     183.0 ± 0%   177.0 ± 0%  -3.28% (p=0.000 n=10)
Unmarshal/untyped_stdlib-10          534.0 ± 0%   534.0 ± 0%       ~ (p=1.000 n=10) ¹
Unmarshal/untyped_unmarshal-10       546.0 ± 0%   538.0 ± 0%  -1.47% (p=0.000 n=10)
Unmarshal/untyped_strict-10          560.0 ± 0%   552.0 ± 0%  -1.43% (p=0.000 n=10)
Unmarshal/untyped_strict-custom-10   561.0 ± 0%   553.0 ± 0%  -1.43% (p=0.000 n=10)
geomean                              310.8        305.2       -1.81%

```