@jpbetz 2024-10-10
/lgtm
/approve
Thanks for getting to the bottom of what changed and how it behaves.